requirements.txt
aiohttp
discord.py==2.3.1
jishaku
python-dotenv
pydash